Includes are changed and cntnfa is now separated from nfa
