+++ include/mata/cntnfa/delta.hh
@@ -6,13 +6,13 @@
 // #include "mata/utils/sparse-set.hh"
 // #include "mata/utils/synchronized-iterator.hh"
 // #include "mata/alphabet.hh"
-// #include "mata/nfa/types.hh"
+// #include "types.hh"
 
 // Note: Temporary local includes for convenience.
 #include "../utils/sparse-set.hh"
 #include "../utils/synchronized-iterator.hh"
 #include "../alphabet.hh"
-#include "../nfa/types.hh"
+#include "types.hh"
 
 #include <iterator>
